docs: installation: improve read- and understandabilty
This patch contains the larger changes, such as splitting longer
sentences and rewording some parts.

The headline for "Reuse a Proxmox Repository for Offline Mirror Tool"
has been changed to "Set up the Repository on non Proxmox bases
systems". Sounds a bit nicer and we are in the Proxmox Offline Mirror
docs anyway.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>